Fix broken sound engine
