+++ src/main/java/org/orecruncher/dsurround/mixins/audio/MixinSoundEngine.java
@@ -1,70 +1,91 @@
 package org.orecruncher.dsurround.mixins.audio;
 
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.mojang.blaze3d.audio.Library;
-import org.lwjgl.openal.ALC10;
-import org.lwjgl.openal.EXTEfx;
-import org.lwjgl.openal.SOFTOutputLimiter;
+import net.minecraft.client.resources.sounds.Sound;
+import net.minecraft.client.resources.sounds.SoundInstance;
+import net.minecraft.client.sounds.ChannelAccess;
+import net.minecraft.client.sounds.SoundEngine;
+import net.minecraft.client.sounds.WeighedSoundEvents;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.phys.Vec3;
 import org.orecruncher.dsurround.config.Configuration;
-import org.orecruncher.dsurround.lib.di.ContainerManager;
+import org.orecruncher.dsurround.mixinutils.MixinHelpers;
 import org.orecruncher.dsurround.runtime.audio.AudioUtilities;
-import org.orecruncher.dsurround.mixinutils.ISoundEngine;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
+import org.orecruncher.dsurround.runtime.audio.SoundFXProcessor;
+import org.orecruncher.dsurround.sound.SoundInstanceHandler;
+import org.orecruncher.dsurround.sound.SoundVolumeEvaluator;
+import org.spongepowered.asm.mixin.*;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
-import java.nio.IntBuffer;
+import java.util.concurrent.CompletableFuture;
 
-@Mixin(Library.class)
-public class MixinSoundEngine implements ISoundEngine {
+@Mixin(SoundEngine.class)
+public class MixinSoundEngine {
 
+    @Final
     @Shadow
-    private long currentDevice;
+    private Library library;
 
-    public long dsurround_getDevicePointer() {
-        return this.currentDevice;
+    @Inject(method = "loadLibrary()V", at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/audio/Library;init(Ljava/lang/String;Z)V", shift = At.Shift.AFTER))
+    public void dsurround_init(CallbackInfo ci) {
+        AudioUtilities.initialize(this.library);
+    }
+
+    @Inject(method = "destroy()V", at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/audio/Library;cleanup()V", shift = At.Shift.BEFORE))
+    public void dsurround_deinit(CallbackInfo ci) {
+        AudioUtilities.deinitialize(this.library);
     }
 
     /**
-     * This will resize the capability buffer to accommodate additional settings
+     * Callback will trigger the creation of sound context information for the sound play once it has been queued to the
+     * sound engine.  It will also perform the first calculations of sound effects based on the player environment.
      */
-    @ModifyConstant(method = "init(Ljava/lang/String;Z)V", constant = @Constant(intValue = 3))
-    private int dsurround_modifyIntBufferSize(int size) {
-        return AudioUtilities.doEnhancedSounds() ? 5 : 3;
+    @Inject(method = "play(Lnet/minecraft/client/resources/sounds/SoundInstance;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/sounds/ChannelAccess$ChannelHandle;execute(Ljava/util/function/Consumer;)V", shift = At.Shift.AFTER), locals = LocalCapture.CAPTURE_FAILEXCEPTION)
+    public void dsurround_onSoundPlay(SoundInstance soundInstance, CallbackInfo ci, WeighedSoundEvents weighedSoundEvents, ResourceLocation resourceLocation, Sound sound, float f, float g, SoundSource soundSource, float h, float i, SoundInstance.Attenuation attenuation, boolean bl, Vec3 vec3, boolean bl2, boolean bl3, CompletableFuture completableFuture, ChannelAccess.ChannelHandle channelHandle) {
+        try {
+            SoundFXProcessor.onSoundPlay(soundInstance, channelHandle);
+            AudioUtilities.onSoundPlay(soundInstance);
+        } catch(final Throwable t) {
+            MixinHelpers.LOGGER.error(t, "Error in dsurround_onSoundPlay()!");
+        }
     }
 
-    /**
-     * Rewrite the capability buffer.  We only do this if advanced processing is enabled.
-     */
-    @WrapOperation(method = "init(Ljava/lang/String;Z)V", at = @At(value = "INVOKE", target = "Lorg/lwjgl/openal/ALC10;alcCreateContext(JLjava/nio/IntBuffer;)J", remap = false))
-    private long dsurround_buildCapabilities(long deviceHandle, IntBuffer attrList, Operation<Long> original) {
-        if (AudioUtilities.doEnhancedSounds()) {
-            // Buffer should have been resized by the constant modification above
-            attrList.clear();
-            // From the original code
-            attrList.put(SOFTOutputLimiter.ALC_OUTPUT_LIMITER_SOFT).put(ALC10.ALC_TRUE);
-            // Increase sends channels
-            attrList.put(EXTEfx.ALC_MAX_AUXILIARY_SENDS).put(4);
-            // Done!
-            attrList.put(0);
-            attrList.flip();
-            return ALC10.alcCreateContext(deviceHandle, attrList);
-        } else {
-            return original.call(deviceHandle, attrList);
+    @Inject(method = "play(Lnet/minecraft/client/resources/sounds/SoundInstance;)V", at = @At("HEAD"), cancellable = true)
+    private void dsurround_play(SoundInstance sound, CallbackInfo ci) {
+        try {
+            if (SoundInstanceHandler.shouldBlockSoundPlay(sound))
+                ci.cancel();
+        } catch (final Exception ignore) {
         }
     }
 
     /**
-     * Modify the number of streaming sounds that can be handled by the underlying sound engine.  The number of
-     * channels to set is driven by config settings.
-     *
-     * @param v Existing value for the number of streaming sounds (should be 8)
-     * @return The quantity of streaming sounds (should be at least 8)
+     * @author
+     * @reason Replacing algorithm for adjusting volume using other external factors
      */
-    @ModifyConstant(method = "init(Ljava/lang/String;Z)V", constant = @Constant(intValue = 8))
-    public int dsurround_initialize(int v) {
-        var config = ContainerManager.resolve(Configuration.SoundSystem.class);
-        return config.streamingChannels;
+    @Overwrite()
+    private float calculateVolume(SoundInstance sound) {
+        return SoundVolumeEvaluator.getAdjustedVolume(sound);
+    }
+
+    @Redirect(method = "play(Lnet/minecraft/client/resources/sounds/SoundInstance;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/sounds/SoundEngine;calculateVolume(FLnet/minecraft/sounds/SoundSource;)F"))
+    private float dsurround_playGetAdjustedVolume(SoundEngine instance, float f, SoundSource soundSource, SoundInstance sound) {
+        return SoundVolumeEvaluator.getAdjustedVolume(sound);
+    }
+
+    @Inject(method = "play(Lnet/minecraft/client/resources/sounds/SoundInstance;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/phys/Vec3;<init>(DDD)V"), locals = LocalCapture.CAPTURE_FAILEXCEPTION, cancellable = true)
+    private void dsurround_soundRangeCheck(SoundInstance soundInstance, CallbackInfo ci, WeighedSoundEvents weighedSoundEvents, ResourceLocation resourceLocation, Sound sound, float f, float g, SoundSource soundSource, float h, float i, SoundInstance.Attenuation attenuation, boolean bl) {
+        if (MixinHelpers.soundSystemConfig.enableSoundPruning) {
+            // If not in range of the listener, cancel.
+            if (!SoundInstanceHandler.inRange(AudioUtilities.getSoundListener().getTransform().position(), soundInstance, 4)) {
+                MixinHelpers.LOGGER.debug(Configuration.Flags.BASIC_SOUND_PLAY, () -> "TOO FAR: " + AudioUtilities.debugString(soundInstance));
+                ci.cancel();
+            }
+        }
     }
 }
