+++ src/main/resources/data/dsurround/tags/blocks/reflectance/high.json
@@ -20,6 +20,17 @@
     "minecraft:deepslate_brick_wall",
     "minecraft:deepslate_tile_slab",
     "minecraft:deepslate_tile_stairs",
-    "minecraft:deepslate_tile_wall"
+    "minecraft:deepslate_tile_wall",
+    "minecraft:cobbled_deepslate",
+    "minecraft:cobbled_deepslate_slab",
+    "minecraft:cobbled_deepslate_stairs",
+    "minecraft:cobblestone",
+    "minecraft:cobblestone_slab",
+    "minecraft:cobblestone_stairs",
+    "minecraft:copper_block",
+    "minecraft:nether_brick_slab",
+    "minecraft:nether_brick_stairs",
+    "minecraft:nether_bricks",
+    "minecraft:basalt"
   ]
 }
