+++ src/main/resources/data/dsurround/tags/blocks/reflectance/medium.json
@@ -6,6 +6,7 @@
     },
     "#minecraft:terracotta",
     "minecraft:dripstone_block",
-    "minecraft:pointed_dripstone"
+    "minecraft:pointed_dripstone",
+    "minecraft:copper_door"
   ]
 }
