+++ src/main/resources/dsurround.mixins.json
@@ -6,14 +6,14 @@
   "mixins": [
   ],
   "client": [
-    "audio.MixinSoundEngine",
+    "audio.MixinSoundLibrary",
     "audio.MixinSoundEvent",
     "audio.MixinSoundManagerAccessor",
-    "audio.MixinSoundSystem",
-    "audio.MixinSoundSystemAccessors",
+    "audio.MixinSoundEngine",
+    "audio.MixinSoundEngineAccessor",
     "audio.MixinSource",
-    "audio.MixinSourceManagerAccessor",
-    "audio.MixinStaticSoundAccessor",
+    "audio.MixinChannelHandleAccessor",
+    "audio.MixinSoundBuffer",
     "core.MixinAbstractSoundInstance",
     "core.MixinBiome",
     "core.MixinBlockState",
